requirements.txt
requests
Flask
python-dateutil
ccxt
